+++ projects/igniteui-angular/src/lib/grids/columns/column.component.ts
@@ -42,7 +42,10 @@ import {
     IgxCollapsibleIndicatorTemplateDirective,
     IgxFilterCellTemplateDirective
 } from './templates.directive';
-import { MRLResizeColumnInfo, MRLColumnSizeInfo } from './interfaces';
+import { MRLResizeColumnInfo, MRLColumnSizeInfo, IColumnPipeArgs } from './interfaces';
+
+const DEFAULT_DATE_FORMAT = 'mediumDate';
+const DEFAULT_DIGITS_INFO = '1.0-3';
 
 /**
  * **Ignite UI for Angular Column** -
@@ -1179,6 +1182,43 @@ export class IgxColumnComponent implements AfterContentInit {
     get visibleWhenCollapsed(): boolean {
         return this._visibleWhenCollapsed;
     }
+
+    private _pipeArgs: IColumnPipeArgs = { format: DEFAULT_DATE_FORMAT, digitsInfo: DEFAULT_DIGITS_INFO };
+    /**
+     * @remarks
+     * Provide parameters for DatePipe and DecimalPipe to customize the display format for date and numeric columns.
+     * Accepts an `IColumnPipeArgs` object with any of the `format`, `timezone` and `digitsInfo` properties.
+     * For more details see https://angular.io/api/common/DatePipe and https://angular.io/api/common/DecimalPipe
+     * @param IColumnPipeArgs object.
+     *
+     * @memberof IgxColumnComponent
+     */
+    @Input()
+    /**
+     * Sets the _pipeArgs.
+     * @example
+     * ```typescript
+     * const formatOptions: IColumnPipeArgs = {
+     *      format: 'longDate',
+     *      timezone: 'UTC'
+     * }
+     * ```
+     * ```html
+     * <igx-column dataType="date" pipeArgs="formatOptions"></igx-column>
+     * <igx-column dataType="number" pipeArgs="{ digitsInfo: '1.1-2' }"></igx-column>
+     * ```
+     */
+    set pipeArgs(value: IColumnPipeArgs) {
+        this._pipeArgs = Object.assign(this._pipeArgs, value);
+        this.grid.summaryService.clearSummaryCache();
+        (this.grid as any)._pipeTrigger++;
+        this.grid.summaryService.retriggerRootPipe++;
+        this.grid.notifyChanges(true);
+    }
+    get pipeArgs(): IColumnPipeArgs {
+        return this._pipeArgs;
+    }
+
 
     /**
      * @hidden