+++ projects/igniteui-angular/src/lib/grids/grid/grid.component.spec.ts
@@ -1456,16 +1456,25 @@ describe('IgxGrid Component Tests #grid', () => {
             });
 
             grid.locale = 'de-DE';
-            tick(500);
-            fixture.detectChanges();
+            grid.columnList.toArray()[4].pipeArgs = {
+                timezone: 'UTC',
+                format: 'longDate',
+                digitsInfo: '1.2-2'
+            };
+            grid.columnList.toArray()[3].pipeArgs = {
+                timezone: 'UTC',
+                format: 'longDate',
+                digitsInfo: '1.2-2'
+            };
+            tick(300);
             fixture.detectChanges();
 
             rows = grid.rowList.toArray();
-            expectedValue = '21.03.2005';
+            expectedValue = '21. März 2005';
             expect(rows[0].cells.toArray()[4].element.nativeElement.textContent).toBe(expectedValue);
-            expectedValue = '15.01.2008';
+            expectedValue = '5. Januar 2008';
             expect(rows[1].cells.toArray()[4].element.nativeElement.textContent).toBe(expectedValue);
-            expectedValue = '20.11.2010';
+            expectedValue = '20. November 2010';
             expect(rows[2].cells.toArray()[4].element.nativeElement.textContent).toBe(expectedValue);
 
             // verify summaries formatting
@@ -1475,11 +1484,11 @@ describe('IgxGrid Component Tests #grid', () => {
                 const earliest = summary.query(By.css('[title=\'Earliest\']'));
                 if (avgLabel) {
                     avgValue = avgLabel.nativeElement.nextSibling.innerText;
-                    expect(avgValue).toBe('3.900,4');
+                    expect(avgValue).toBe('3.900,40');
                 }
                 if (earliest) {
                     earliestValue = earliest.nativeElement.nextSibling.innerText;
-                    expect(earliestValue).toBe('17.05.1990');
+                    expect(earliestValue).toBe('17. Mai 1990');
                 }
             });
         }));